# Taylor approximations (or other approximation) for LOG2/EXP2/SIN in function.py passing all tests