Here is my approach for the approximation of SIN, LOG2, EXP2 (which i believe has been provided by 2^x). I've relied on the following resources:
 - https://www.netlib.org/fdlibm/
 
For tests and benchmarks I've used the following environment:
### Environment:
- Machine
```
NVIDIA Corporation GM108M [GeForce 840M]
Intel(R) Core(TM) i7-5600U CPU @ 2.60GHz, CPU(s): 4, Thread(s) per core: 2
Core(s) per socket: 2
Socket(s): 1
```
- Software 
```
OS: Ubuntu 22.04
Kernel: x86_64 Linux 6.5.0-26-generic
Python 3.10.12
torch==2.3.1+cu121
```
### Tests
I ran test for the whole suite except for `test_real_world.py` because I suspect I don't have enough compute:
```
.Killed
```
The other error is from the file `test_dtype.py`, even the master branch fails on this file:
```
Ran 174 tests in 35.090s
FAILED (failures=2, skipped=12)
File "/home/humbulani/tinygrad/tinygrad/dtype.py", line 87, in <module>
    dtypes.default_float = getattr(dtypes, env_default_float.lower())
AttributeError: type object 'dtypes' has no attribute 'typo'
File "/home/humbulani/tinygrad_env/lib/python3.10/site-packages/numpy/testing/_private/utils.py", line 797, in assert_array_compare
    raise AssertionError(msg)
AssertionError: 
Not equal to tolerance rtol=0.001, atol=0
```
### Benchmarks
- benchmarks before:
```
UOPS_REWRITE=1 PROFILE=0 python test/external/external_benchmark_schedule.py
***** model forward in  69.64 ms
***** model schedule in  21.78 ms
***** model lower in 2753.58 ms
```
- benchmarks after:
```
UOPS_REWRITE=1 PROFILE=0 python test/external/external_benchmark_schedule.py
***** model forward in  68.37 ms
***** model schedule in  20.93 ms
***** model lower in 2652.74 ms
```